+++ desk/src/lib/controls/Control.svelte
@@ -84,10 +84,13 @@
 {/snippet}
 
 {#if ft === "Check"}
-  <div class="field check" class:compact>
-    <input {id} type="checkbox" checked={!!value} disabled={ro} onchange={(e) => onchange((e.target as HTMLInputElement).checked)} />
-    {#if !inGrid}<label for={id}>{field.label}</label>{/if}
-    {@render fieldButtons()}
+  <div class="field check" class:compact class:bold={field.bold}>
+    <div class="check-control">
+      <input {id} type="checkbox" checked={!!value} disabled={ro} onchange={(e) => onchange((e.target as HTMLInputElement).checked)} />
+      {#if !inGrid}<label for={id}>{field.label}{#if req && !ro}<span class="req">*</span>{/if}</label>{/if}
+      {@render fieldButtons()}
+    </div>
+    {#if shownError}<div class="err">{shownError}</div>{:else if field.description && !inGrid}<div class="desc">{field.description}</div>{/if}
   </div>
 {:else}
   <div class="field" class:compact class:bold={field.bold}>